requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify ~= 1.7.0
playwright ~= 1.39.0
random-user-agent==1.0.1
httpx==0.27.0
bs4==0.0.2
python-dotenv==1.0.1